Merge r580 from the LLVM vendor branch removing quadratic space use in long functions.  This reduces test_compile's memory use from ~8GB (reported by Thomas on a 64-bit build) to ~987MB (reported by /usr/bin/time -l on OS X for a 32-bit build).  test_compile still takes 45 minutes to run. ------------------------------------------------------------------------ r72411 | evancheng | 2009-05-25 23:25:46 -0700 (Mon, 25 May 2009) | 1 line
Eliminate VarInfo::UsedBlocks.
------------------------------------------------------------------------
------------------------------------------------------------------------
r72426 | jyasskin | 2009-05-26 11:27:15 -0700 (Tue, 26 May 2009) | 7 lines

LiveVariables::VarInfo contains an AliveBlocks BitVector, which has as many
entries as there are basic blocks in the function.  LiveVariables::getVarInfo
creates a VarInfo struct for every register in the function, leading to
quadratic space use.  This patch changes the BitVector to a SparseBitVector,
which doesn't help the worst-case memory use but does reduce the actual use in
very long functions with short-lived variables.

------------------------------------------------------------------------



git-svn-id: http://unladen-swallow.googlecode.com/svn/trunk@583 05521daa-c0b4-11dd-bb00-bd6ab96fe29a